refactor, move main.go under cmd
